Refactored test to IT since it is dependent on Enhetsregistreret
